Extract Remaining 'other_tags' as Comma-Separated Strings

Parses the 'other_tags' column from an imported OSM PBF file and reformats it as a comma-separated key:"value" string per feature, optionally dropping a set of tag keys. The intended use is to preserve OSM tags not already captured by osm_classify(): pass the classification's tag keys via exclude so they are not duplicated in main_tag, main_tag_value, and alt_tags_values.

Usage

osm_other_tags_str(x, exclude = character(0), split = "\",\"|\"=>\"", ...)

Arguments

x

character. The 'other_tags' column of an imported osm.pbf file.

exclude

character. Tag keys to drop from the output. Default character(0).

split

character. Pattern passed to strsplit to split up x.

...

further arguments to strsplit.

Value

a character vector of the same length as x, with each element a comma-separated string of key:"value" pairs, or NA_character_ if no tags remain after exclusion (or the input was missing/empty).

See also

Examples

# See Examples at ?osmclass for full examples

# Reformat all other_tags as key:"value" strings
head(osm_other_tags_str(djibouti_points$other_tags))
#> [1] "GNS:dsg_code:\"PPLC\", GNS:dsg_name:\"populated place\", GNS:id:\"-2034412\", GNS:modify_date:\"2018-04-11\", admin_level:\"2\", alt_name:ar:\"دجبت;غإبت;جبت;جبوت;جيبوتي;جيبوتى;دجيبوتي;مدينة جيبوتي\", capital:\"yes\", capital_ISO3166-1:\"yes\", is_capital:\"country\", name:ace:\"Djibouti\", name:af:\"Djiboeti\", name:am:\"ጅቡቲ\", name:an:\"Chibuti\", name:ar:\"جيبوتي\", name:arz:\"جيبوتي\", name:ast:\"Xibuti\", name:az:\"Cibuti\", name:azb:\"جیبوتی\", name:ba:\"Джибути\", name:bat-smg:\"Džėbotis\", name:bcl:\"Dibouti\", name:be:\"Джыбуці\", name:bg:\"Джибути\", name:bjn:\"Djibouti\", name:bn:\"জিবুতি\", name:bo:\"ཇི་བའོ་ཊི།\", name:bpy:\"ডিজিবোটি\", name:br:\"Djibouti\", name:bs:\"Djibouti\", name:bxr:\"Джибути\", name:ca:\"Djibouti\", name:ce:\"Джибути\", name:ceb:\"Yibuti\", name:ckb:\"جیبووتی\", name:crh:\"Cibuti\", name:cs:\"Džíbutí\", name:cv:\"Джибути\", name:cy:\"Djibouti\", name:da:\"Djibouti\", name:de:\"Dschibuti\", name:diq:\"Cibuti\", name:dv:\"ޖިބުތީ\", name:ee:\"Djibouti\", name:el:\"Τζιμπουτί\", name:en:\"Djibouti\", name:eo:\"Ĝibutio\", name:es:\"Yibuti\", name:et:\"Djibouti\", name:eu:\"Djibuti\", name:ext:\"Yibuti\", name:fa:\"جیبوتی\", name:fi:\"Djibouti\", name:fr:\"Djibouti\", name:frp:\"Dj·iboti\", name:fy:\"Dzjibûty\", name:ga:\"Diobúití\", name:gag:\"Cibuti\", name:gd:\"Djibouti\", name:gl:\"Xibutí\", name:gv:\"Djibouti\", name:he:\"ג'יבוטי\", name:hi:\"जीबूती\", name:hif:\"Djibouti\", name:hr:\"Djibouti\", name:hsb:\"Dźibuti\", name:ht:\"Djibouti\", name:hu:\"Dzsibuti\", name:hy:\"Ջիբութի\", name:ia:\"Djibouti\", name:id:\"Djibouti\", name:ie:\"Djibouti\", name:ilo:\"Djibouti\", name:io:\"Djibuti\", name:is:\"Djíbútí\", name:it:\"Gibuti\", name:ja:\"ジブチ市\", name:jv:\"Djibouti\", name:ka:\"ჯიბუტი\", name:kab:\"Jibuti\", name:kab-Arab:\"جيبوتي\", name:kg:\"Djibuti\", name:ki:\"Djibouti\", name:kk:\"Джибути\", name:kk-Arab:\"دجىيبۋتىي\", name:kmr:\"جیبووتی\", name:kn:\"ಜಿಬೂತೀ\", name:ko:\"지부티\", name:ks:\"جیبوتی\", name:ku:\"Cîbûtî\", name:ku-Arab:\"جیبووتی\", name:kw:\"Sita Jibouti\", name:ky:\"Жибути шаары\", name:la:\"Urbs Dzibutum\", name:lb:\"Dschibuti\", name:li:\"Djibouti\", name:lij:\"Gibuti\", name:lmo:\"Djibouti\", name:ln:\"Djibuti\", name:lrc:\"جیبۊتی\", name:lt:\"Džibutis\", name:lv:\"Džibuti\", name:lzh:\"吉布地\", name:mg:\"Jibotia\", name:min:\"Djibouti\", name:mk:\"Џибути\", name:ml:\"ജിബൂട്ടി\", name:mr:\"जिबूती\", name:mrj:\"Джибути\", name:ms:\"Djibouti\", name:mt:\"Ġibuti\", name:my:\"ဂျီဘူတီနိုင်ငံ\", name:mzn:\"جیبوتی\", name:na:\"Djibuti\", name:nah:\"Yibuti\", name:nan:\"Djibouti\", name:nds:\"Dschibuti\", name:ne:\"जिबुटी\", name:nl:\"Djibouti\", name:nn:\"Djibouti\", name:no:\"Djibouti\", name:nov:\"Djibuti\", name:nso:\"Djibouti\", name:nv:\"Jibótii\", name:oc:\"Jiboti\", name:om:\"Jibuutii\", name:or:\"ଡିଜିବୋଇଟି\", name:os:\"Джибути\", name:pa:\"ਜਿਬੂਤੀ\", name:pam:\"Djibouti\", name:pdc:\"Tschibuti\", name:pih:\"Jibuuti\", name:pl:\"Dżibuti\", name:pms:\"Djibouti\", name:pnb:\"جبوتی\", name:ps:\"جېبوتي\", name:pt:\"Djibouti\", name:qu:\"Yiwuti\", name:ro:\"Djibouti\", name:ru:\"Джибути\", name:rue:\"Джібутьско\", name:rw:\"Jibuti\", name:sa:\"जिबूटी\", name:sah:\"Дьибути\", name:sc:\"Gibuti\", name:scn:\"Gibbuti\", name:sco:\"Djibouti\", name:sd:\"جبوتي\", name:se:\"Djibouti\", name:sg:\"Dibutùii\", name:sh:\"Djibouti\", name:simple:\"Djibouti\", name:sk:\"Džibuti\", name:sl:\"Džibuti\", name:sn:\"Djibouti\", name:so:\"Jabuuti\", name:sq:\"Xhibuti\", name:sr:\"Џибути\", name:ss:\"IJibhuthi\", name:stq:\"Dschibuti\", name:su:\"Djibouti\", name:sv:\"Djibouti\", name:sw:\"Jibuti\", name:ta:\"சீபூத்தீ\", name:te:\"జిబౌటి\", name:tg:\"Ҷибути\", name:th:\"จิบูตี\", name:ti:\"ጂቡቲ\", name:tk:\"Jibuti\", name:tl:\"Djibouti\", name:tr:\"Cibuti\", name:ts:\"Djibouti\", name:tt:\"Җибути\", name:udm:\"Джибути\", name:ug:\"جىبۇتى\", name:uk:\"Джибуті\", name:ur:\"جبوتی\", name:uz:\"Jibuti\", name:vec:\"Gibuti\", name:vi:\"Djibouti\", name:vo:\"Cibutän\", name:vro:\"Djibouti\", name:war:\"Djibouti\", name:wo:\"Jibuti\", name:xal:\"Җибудин\", name:yi:\"דזשיבוטי\", name:yo:\"Djibouti\", name:yue:\"吉布提\", name:zh:\"吉布地市\", name:zu:\"IJibuthi\", population:\"475332\", wikidata:\"Q3604\", wikipedia:\"ar:جيبوتي (مدينة)\", wikipedia:ar:\"جيبوتي (مدينة)\"\""
#> [2] "traffic_signals:\"signal\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [3] NA                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [4] NA                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [5] NA                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [6] "noexit:\"yes\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   

# Drop tag keys used by the classification (already captured in
# 'main_tag'/'alt_tags_values' returned by osm_classify())
used_tags <- unique(unlist(lapply(osm_point_polygon_class, names)))
head(osm_other_tags_str(djibouti_points$other_tags, exclude = used_tags))
#> [1] "GNS:dsg_code:\"PPLC\", GNS:dsg_name:\"populated place\", GNS:id:\"-2034412\", GNS:modify_date:\"2018-04-11\", admin_level:\"2\", alt_name:ar:\"دجبت;غإبت;جبت;جبوت;جيبوتي;جيبوتى;دجيبوتي;مدينة جيبوتي\", capital:\"yes\", capital_ISO3166-1:\"yes\", is_capital:\"country\", name:ace:\"Djibouti\", name:af:\"Djiboeti\", name:am:\"ጅቡቲ\", name:an:\"Chibuti\", name:ar:\"جيبوتي\", name:arz:\"جيبوتي\", name:ast:\"Xibuti\", name:az:\"Cibuti\", name:azb:\"جیبوتی\", name:ba:\"Джибути\", name:bat-smg:\"Džėbotis\", name:bcl:\"Dibouti\", name:be:\"Джыбуці\", name:bg:\"Джибути\", name:bjn:\"Djibouti\", name:bn:\"জিবুতি\", name:bo:\"ཇི་བའོ་ཊི།\", name:bpy:\"ডিজিবোটি\", name:br:\"Djibouti\", name:bs:\"Djibouti\", name:bxr:\"Джибути\", name:ca:\"Djibouti\", name:ce:\"Джибути\", name:ceb:\"Yibuti\", name:ckb:\"جیبووتی\", name:crh:\"Cibuti\", name:cs:\"Džíbutí\", name:cv:\"Джибути\", name:cy:\"Djibouti\", name:da:\"Djibouti\", name:de:\"Dschibuti\", name:diq:\"Cibuti\", name:dv:\"ޖިބުތީ\", name:ee:\"Djibouti\", name:el:\"Τζιμπουτί\", name:en:\"Djibouti\", name:eo:\"Ĝibutio\", name:es:\"Yibuti\", name:et:\"Djibouti\", name:eu:\"Djibuti\", name:ext:\"Yibuti\", name:fa:\"جیبوتی\", name:fi:\"Djibouti\", name:fr:\"Djibouti\", name:frp:\"Dj·iboti\", name:fy:\"Dzjibûty\", name:ga:\"Diobúití\", name:gag:\"Cibuti\", name:gd:\"Djibouti\", name:gl:\"Xibutí\", name:gv:\"Djibouti\", name:he:\"ג'יבוטי\", name:hi:\"जीबूती\", name:hif:\"Djibouti\", name:hr:\"Djibouti\", name:hsb:\"Dźibuti\", name:ht:\"Djibouti\", name:hu:\"Dzsibuti\", name:hy:\"Ջիբութի\", name:ia:\"Djibouti\", name:id:\"Djibouti\", name:ie:\"Djibouti\", name:ilo:\"Djibouti\", name:io:\"Djibuti\", name:is:\"Djíbútí\", name:it:\"Gibuti\", name:ja:\"ジブチ市\", name:jv:\"Djibouti\", name:ka:\"ჯიბუტი\", name:kab:\"Jibuti\", name:kab-Arab:\"جيبوتي\", name:kg:\"Djibuti\", name:ki:\"Djibouti\", name:kk:\"Джибути\", name:kk-Arab:\"دجىيبۋتىي\", name:kmr:\"جیبووتی\", name:kn:\"ಜಿಬೂತೀ\", name:ko:\"지부티\", name:ks:\"جیبوتی\", name:ku:\"Cîbûtî\", name:ku-Arab:\"جیبووتی\", name:kw:\"Sita Jibouti\", name:ky:\"Жибути шаары\", name:la:\"Urbs Dzibutum\", name:lb:\"Dschibuti\", name:li:\"Djibouti\", name:lij:\"Gibuti\", name:lmo:\"Djibouti\", name:ln:\"Djibuti\", name:lrc:\"جیبۊتی\", name:lt:\"Džibutis\", name:lv:\"Džibuti\", name:lzh:\"吉布地\", name:mg:\"Jibotia\", name:min:\"Djibouti\", name:mk:\"Џибути\", name:ml:\"ജിബൂട്ടി\", name:mr:\"जिबूती\", name:mrj:\"Джибути\", name:ms:\"Djibouti\", name:mt:\"Ġibuti\", name:my:\"ဂျီဘူတီနိုင်ငံ\", name:mzn:\"جیبوتی\", name:na:\"Djibuti\", name:nah:\"Yibuti\", name:nan:\"Djibouti\", name:nds:\"Dschibuti\", name:ne:\"जिबुटी\", name:nl:\"Djibouti\", name:nn:\"Djibouti\", name:no:\"Djibouti\", name:nov:\"Djibuti\", name:nso:\"Djibouti\", name:nv:\"Jibótii\", name:oc:\"Jiboti\", name:om:\"Jibuutii\", name:or:\"ଡିଜିବୋଇଟି\", name:os:\"Джибути\", name:pa:\"ਜਿਬੂਤੀ\", name:pam:\"Djibouti\", name:pdc:\"Tschibuti\", name:pih:\"Jibuuti\", name:pl:\"Dżibuti\", name:pms:\"Djibouti\", name:pnb:\"جبوتی\", name:ps:\"جېبوتي\", name:pt:\"Djibouti\", name:qu:\"Yiwuti\", name:ro:\"Djibouti\", name:ru:\"Джибути\", name:rue:\"Джібутьско\", name:rw:\"Jibuti\", name:sa:\"जिबूटी\", name:sah:\"Дьибути\", name:sc:\"Gibuti\", name:scn:\"Gibbuti\", name:sco:\"Djibouti\", name:sd:\"جبوتي\", name:se:\"Djibouti\", name:sg:\"Dibutùii\", name:sh:\"Djibouti\", name:simple:\"Djibouti\", name:sk:\"Džibuti\", name:sl:\"Džibuti\", name:sn:\"Djibouti\", name:so:\"Jabuuti\", name:sq:\"Xhibuti\", name:sr:\"Џибути\", name:ss:\"IJibhuthi\", name:stq:\"Dschibuti\", name:su:\"Djibouti\", name:sv:\"Djibouti\", name:sw:\"Jibuti\", name:ta:\"சீபூத்தீ\", name:te:\"జిబౌటి\", name:tg:\"Ҷибути\", name:th:\"จิบูตี\", name:ti:\"ጂቡቲ\", name:tk:\"Jibuti\", name:tl:\"Djibouti\", name:tr:\"Cibuti\", name:ts:\"Djibouti\", name:tt:\"Җибути\", name:udm:\"Джибути\", name:ug:\"جىبۇتى\", name:uk:\"Джибуті\", name:ur:\"جبوتی\", name:uz:\"Jibuti\", name:vec:\"Gibuti\", name:vi:\"Djibouti\", name:vo:\"Cibutän\", name:vro:\"Djibouti\", name:war:\"Djibouti\", name:wo:\"Jibuti\", name:xal:\"Җибудин\", name:yi:\"דזשיבוטי\", name:yo:\"Djibouti\", name:yue:\"吉布提\", name:zh:\"吉布地市\", name:zu:\"IJibuthi\", population:\"475332\", wikidata:\"Q3604\", wikipedia:\"ar:جيبوتي (مدينة)\", wikipedia:ar:\"جيبوتي (مدينة)\"\""
#> [2] "traffic_signals:\"signal\""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> [3] NA                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [4] NA                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [5] NA                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [6] "noexit:\"yes\""