A Classification of OSM Features by Economic Function

This classification, developed for Krantz (2023), aims to classify OSM features into meaningful and specific economic categories such as 'education', 'health', 'tourism', 'financial', 'shopping', 'transport', 'communications', 'industrial', 'residential', 'road', 'railway', 'pipeline', 'power', 'waterway' etc. Separate classifications are developed for points and polygons (buildings) (33 categories), and lines (11 categories), which should be applied to the respective layers of OSM PBF files, see osmclass-package for and example. The classification is optimized (in terms of tag choice and order of categories) to assign the most sensible primary category to most features in the Africa OSM.

Usage

osm_point_polygon_class

osm_line_class

osm_line_info_tags

Format

An object of class list of length 33.

An object of class list of length 11.

An object of class list of length 11.

References

Krantz, Sebastian, Mapping Africa’s Infrastructure Potential with Geospatial Big Data, Causal ML, and XAI (August 10, 2023). Available at SSRN: https://www.ssrn.com/abstract=4537867

See also

Examples

collapse::unlist2d(osm_point_polygon_class, idcols = c("category", "tag"))
#>            category              tag                      V1
#> 1              port          landuse                    port
#> 2          military         military                        
#> 3          military         building                military
#> 4          military          landuse                military
#> 5             craft            craft                        
#> 6         education          amenity                 college
#> 7         education         building                 college
#> 8         education          landuse               education
#> 9     education_alt          amenity          driving_school
#> 10    education_alt           office educational_institution
#> 11    education_alt         building            kindergarten
#> 12           health          amenity              baby_hatch
#> 13           health       healthcare                        
#> 14           health         building                hospital
#> 15          farming            place                    farm
#> 16          farming         man_made                 beehive
#> 17          farming         building                    farm
#> 18          farming          landuse                farmland
#> 19           mining         man_made                    adit
#> 20           mining          landuse                  quarry
#> 21       industrial       industrial                        
#> 22       industrial         man_made                    kiln
#> 23       industrial         building              industrial
#> 24       industrial          landuse              industrial
#> 25    accommodation          tourism               apartment
#> 26    accommodation         building                   hotel
#> 27             food          amenity                     bar
#> 28          tourism          tourism                     !no
#> 29          tourism             shop           travel_agency
#> 30          tourism           office                   guide
#> 31       creativity          amenity       conference_centre
#> 32       creativity          leisure             hackerspace
#> 33       creativity           office          graphic_design
#> 34           sports          leisure          fitness_centre
#> 35           sports            sport                        
#> 36           sports         building              grandstand
#> 37    entertainment          amenity             arts_centre
#> 38    entertainment          leisure     adult_gaming_centre
#> 39    institutional           office              government
#> 40    institutional         building              government
#> 41    institutional          landuse           institutional
#> 42   public_service          amenity              courthouse
#> 43   public_service         building            fire_station
#> 44          storage         man_made                    silo
#> 45          storage         building                  hangar
#> 46          storage          landuse                   depot
#> 47   communications          amenity           internet_cafe
#> 48   communications          telecom                        
#> 49   communications          utility                 telecom
#> 50   communications         man_made                 antenna
#> 51   communications           office       telecommunication
#> 52   communications       tower:type           communication
#> 53   communications    communication                    line
#> 54        transport          amenity         bicycle_parking
#> 55        transport          highway                crossing
#> 56        transport          railway                        
#> 57        transport        aerialway                        
#> 58        transport         waterway              !waterfall
#> 59        transport          aeroway                        
#> 60        transport public_transport                        
#> 61        transport           bridge                        
#> 62        transport         junction                        
#> 63        transport           office               logistics
#> 64        transport         man_made                  bridge
#> 65        transport         building           train_station
#> 66        transport          landuse                 railway
#> 67         shopping          amenity             marketplace
#> 68         shopping             shop                     !no
#> 69         shopping         building                  retail
#> 70         shopping          landuse                  retail
#> 71        financial          amenity                     atm
#> 72        financial           office              accountant
#> 73         religion          amenity            funeral_hall
#> 74         religion         building               cathedral
#> 75         religion           office                religion
#> 76         religion          landuse               religious
#> 77         religion         religion                     !no
#> 78         religion     denomination                        
#> 79     construction         building            construction
#> 80     construction          landuse            construction
#> 81      residential         building                   house
#> 82      residential     building:use             residential
#> 83      residential          landuse             residential
#> 84         historic         historic                     !no
#> 85            waste          amenity   sanitary_dump_station
#> 86            waste            water              wastewater
#> 87            waste         man_made        wastewater_plant
#> 88            waste          landuse                landfill
#> 89     office_other           office                        
#> 90     office_other         building                  office
#> 91       commercial         building              commercial
#> 92       commercial          landuse              commercial
#> 93            power            power                        
#> 94            power          utility                   power
#> 95            power         building       transformer_tower
#> 96            power       tower:type                   power
#> 97  utilities_other         man_made             water_tower
#> 98  utilities_other            water               reservoir
#> 99  utilities_other           office           water_utility
#> 100 utilities_other         building                digester
#> 101 utilities_other          landuse               reservoir
#> 102      recreation          amenity        community_centre
#> 103      recreation          leisure            beach_resort
#> 104      recreation          landuse       recreation_ground
#> 105      recreation         building                   civic
#> 106      facilities          amenity                     bbq
#> 107      facilities         building                 toilets
#> 108       emergency        emergency                     !no
#>                         V2                  V3                   V4
#> 1                     <NA>                <NA>                 <NA>
#> 2                     <NA>                <NA>                 <NA>
#> 3                 barracks              bunker                 <NA>
#> 4                     <NA>                <NA>                 <NA>
#> 5                     <NA>                <NA>                 <NA>
#> 6                   school          university                 <NA>
#> 7                   school          university                 <NA>
#> 8                     <NA>                <NA>                 <NA>
#> 9             kindergarten     language_school              library
#> 10                tutoring                <NA>                 <NA>
#> 11                    <NA>                <NA>                 <NA>
#> 12                  clinic             dentist              doctors
#> 13                    <NA>                <NA>                 <NA>
#> 14                    <NA>                <NA>                 <NA>
#> 15                    <NA>                <NA>                 <NA>
#> 16                    <NA>                <NA>                 <NA>
#> 17                    barn        conservatory              cowshed
#> 18                farmyard             orchard             vineyard
#> 19               mineshaft      petroleum_well        tailings_pond
#> 20                    <NA>                <NA>                 <NA>
#> 21                    <NA>                <NA>                 <NA>
#> 22                   works                <NA>                 <NA>
#> 23                    <NA>                <NA>                 <NA>
#> 24                    <NA>                <NA>                 <NA>
#> 25                  chalet         guest_house               hostel
#> 26                    <NA>                <NA>                 <NA>
#> 27                    cafe           fast_food           food_court
#> 28                    <NA>                <NA>                 <NA>
#> 29                    <NA>                <NA>                 <NA>
#> 30            travel_agent                <NA>                 <NA>
#> 31                  studio                <NA>                 <NA>
#> 32                    <NA>                <NA>                 <NA>
#> 33               coworking            engineer                   it
#> 34         fitness_station               pitch        sports_centre
#> 35                    <NA>                <NA>                 <NA>
#> 36                pavilion         riding_hall          sports_hall
#> 37       exhibition_centre             brothel               casino
#> 38        amusement_arcade           bandstand                dance
#> 39                     ngo         association           diplomatic
#> 40                    <NA>                <NA>                 <NA>
#> 41                    <NA>                <NA>                 <NA>
#> 42            fire_station              police             post_box
#> 43                  public                <NA>                 <NA>
#> 44            storage_tank                <NA>                 <NA>
#> 45            storage_tank                silo                 <NA>
#> 46                    <NA>                <NA>                 <NA>
#> 47               telephone                <NA>                 <NA>
#> 48                    <NA>                <NA>                 <NA>
#> 49                    <NA>                <NA>                 <NA>
#> 50                  beacon communication_tower communications_tower
#> 51                    <NA>                <NA>                 <NA>
#> 52       telecommunication               radar                radio
#> 53                    pole              mobile         mobile_phone
#> 54  bicycle_repair_station      bicycle_rental          boat_rental
#> 55         mini_roundabout   motorway_junction      traffic_signals
#> 56                    <NA>                <NA>                 <NA>
#> 57                    <NA>                <NA>                 <NA>
#> 58                 !stream         !stream_end              !rapids
#> 59                    <NA>                <NA>                 <NA>
#> 60                    <NA>                <NA>                 <NA>
#> 61                    <NA>                <NA>                 <NA>
#> 62                    <NA>                <NA>                 <NA>
#> 63          moving_company                <NA>                 <NA>
#> 64          goods_conveyor                <NA>                 <NA>
#> 65          transportation              bridge                 <NA>
#> 66                    <NA>                <NA>                 <NA>
#> 67                    shop                <NA>                 <NA>
#> 68                 !vacant                <NA>                 <NA>
#> 69                   kiosk         supermarket            warehouse
#> 70                    <NA>                <NA>                 <NA>
#> 71                    bank    bureau_de_change         mobile_money
#> 72               financial   financial_advisor            insurance
#> 73             crematorium          grave_yard            monastery
#> 74                  chapel              church         kingdom_hall
#> 75                    <NA>                <NA>                 <NA>
#> 76                    <NA>                <NA>                 <NA>
#> 77                   !none                <NA>                 <NA>
#> 78                    <NA>                <NA>                 <NA>
#> 79                    <NA>                <NA>                 <NA>
#> 80                    <NA>                <NA>                 <NA>
#> 81              apartments            bungalow             detached
#> 82              apartments                <NA>                 <NA>
#> 83                    <NA>                <NA>                 <NA>
#> 84                    <NA>                <NA>                 <NA>
#> 85               recycling        waste_basket       waste_disposal
#> 86                    <NA>                <NA>                 <NA>
#> 87                    <NA>                <NA>                 <NA>
#> 88                    <NA>                <NA>                 <NA>
#> 89                    <NA>                <NA>                 <NA>
#> 90                    <NA>                <NA>                 <NA>
#> 91                    <NA>                <NA>                 <NA>
#> 92                    <NA>                <NA>                 <NA>
#> 93                    <NA>                <NA>                 <NA>
#> 94                    <NA>                <NA>                 <NA>
#> 95                    <NA>                <NA>                 <NA>
#> 96                    <NA>                <NA>                 <NA>
#> 97              water_well         water_works            water_tap
#> 98                    <NA>                <NA>                 <NA>
#> 99                    <NA>                <NA>                 <NA>
#> 100                service         water_tower                 <NA>
#> 101                   <NA>                <NA>                 <NA>
#> 102               fountain         public_bath      public_bookcase
#> 103                 marina      miniature_golf               common
#> 104                   <NA>                <NA>                 <NA>
#> 105                   <NA>                <NA>                 <NA>
#> 106                  bench          dog_toilet            childcare
#> 107                   <NA>                <NA>                 <NA>
#> 108                   <NA>                <NA>                 <NA>
#>                         V5               V6                      V7
#> 1                     <NA>             <NA>                    <NA>
#> 2                     <NA>             <NA>                    <NA>
#> 3                     <NA>             <NA>                    <NA>
#> 4                     <NA>             <NA>                    <NA>
#> 5                     <NA>             <NA>                    <NA>
#> 6                     <NA>             <NA>                    <NA>
#> 7                     <NA>             <NA>                    <NA>
#> 8                     <NA>             <NA>                    <NA>
#> 9                 training     music_school                    <NA>
#> 10                    <NA>             <NA>                    <NA>
#> 11                    <NA>             <NA>                    <NA>
#> 12                hospital     nursing_home                pharmacy
#> 13                    <NA>             <NA>                    <NA>
#> 14                    <NA>             <NA>                    <NA>
#> 15                    <NA>             <NA>                    <NA>
#> 16                    <NA>             <NA>                    <NA>
#> 17          farm_auxiliary       greenhouse             slurry_tank
#> 18             aquaculture        salt_pond greenhouse_horticulture
#> 19                    <NA>             <NA>                    <NA>
#> 20                    <NA>             <NA>                    <NA>
#> 21                    <NA>             <NA>                    <NA>
#> 22                    <NA>             <NA>                    <NA>
#> 23                    <NA>             <NA>                    <NA>
#> 24                    <NA>             <NA>                    <NA>
#> 25                   hotel            motel          wilderness_hut
#> 26                    <NA>             <NA>                    <NA>
#> 27               ice_cream              pub              restaurant
#> 28                    <NA>             <NA>                    <NA>
#> 29                    <NA>             <NA>                    <NA>
#> 30                    <NA>             <NA>                    <NA>
#> 31                    <NA>             <NA>                    <NA>
#> 32                    <NA>             <NA>                    <NA>
#> 33                research             <NA>                    <NA>
#> 34                 stadium    swimming_pool                   track
#> 35                    <NA>             <NA>                    <NA>
#> 36                 stadium             <NA>                    <NA>
#> 37                  cinema     events_venue                gambling
#> 38                    <NA>             <NA>                    <NA>
#> 39         political_party             <NA>                    <NA>
#> 40                    <NA>             <NA>                    <NA>
#> 41                    <NA>             <NA>                    <NA>
#> 42              post_depot      post_office                  prison
#> 43                    <NA>             <NA>                    <NA>
#> 44                    <NA>             <NA>                    <NA>
#> 45                    <NA>             <NA>                    <NA>
#> 46                    <NA>             <NA>                    <NA>
#> 47                    <NA>             <NA>                    <NA>
#> 48                    <NA>             <NA>                    <NA>
#> 49                    <NA>             <NA>                    <NA>
#> 50              lighthouse   satellite_dish                    <NA>
#> 51                    <NA>             <NA>                    <NA>
#> 52       radio_transmitter             <NA>                    <NA>
#> 53        mobile_phone=yes             <NA>                    <NA>
#> 54            boat_sharing      bus_station           train_station
#> 55          turning_circle     turning_loop                bus_stop
#> 56                    <NA>             <NA>                    <NA>
#> 57                    <NA>             <NA>                    <NA>
#> 58          !tidal_channel       !riverbank                    <NA>
#> 59                    <NA>             <NA>                    <NA>
#> 60                    <NA>             <NA>                    <NA>
#> 61                    <NA>             <NA>                    <NA>
#> 62                    <NA>             <NA>                    <NA>
#> 63                    <NA>             <NA>                    <NA>
#> 64                    <NA>             <NA>                    <NA>
#> 65                    <NA>             <NA>                    <NA>
#> 66                    <NA>             <NA>                    <NA>
#> 67                    <NA>             <NA>                    <NA>
#> 68                    <NA>             <NA>                    <NA>
#> 69                    <NA>             <NA>                    <NA>
#> 70                    <NA>             <NA>                    <NA>
#> 71      mobile_money_agent   money_transfer                    <NA>
#> 72             tax_advisor             <NA>                    <NA>
#> 73       place_of_mourning place_of_worship                    <NA>
#> 74               monastery           mosque              presbytery
#> 75                    <NA>             <NA>                    <NA>
#> 76                    <NA>             <NA>                    <NA>
#> 77                    <NA>             <NA>                    <NA>
#> 78                    <NA>             <NA>                    <NA>
#> 79                    <NA>             <NA>                    <NA>
#> 80                    <NA>             <NA>                    <NA>
#> 81      semidetached_house          terrace               dormitory
#> 82                    <NA>             <NA>                    <NA>
#> 83                    <NA>             <NA>                    <NA>
#> 84                    <NA>             <NA>                    <NA>
#> 85  waste_transfer_station             <NA>                    <NA>
#> 86                    <NA>             <NA>                    <NA>
#> 87                    <NA>             <NA>                    <NA>
#> 88                    <NA>             <NA>                    <NA>
#> 89                    <NA>             <NA>                    <NA>
#> 90                    <NA>             <NA>                    <NA>
#> 91                    <NA>             <NA>                    <NA>
#> 92                    <NA>             <NA>                    <NA>
#> 93                    <NA>             <NA>                    <NA>
#> 94                    <NA>             <NA>                    <NA>
#> 95                    <NA>             <NA>                    <NA>
#> 96                    <NA>             <NA>                    <NA>
#> 97              water_tank        gasometer                pipeline
#> 98                    <NA>             <NA>                    <NA>
#> 99                    <NA>             <NA>                    <NA>
#> 100                   <NA>             <NA>                    <NA>
#> 101                   <NA>             <NA>                    <NA>
#> 102          social_centre             <NA>                    <NA>
#> 103                   park         dog_park              water_park
#> 104                   <NA>             <NA>                    <NA>
#> 105                   <NA>             <NA>                    <NA>
#> 106                  clock      photo_booth                 kitchen
#> 107                   <NA>             <NA>                    <NA>
#> 108                   <NA>             <NA>                    <NA>
#>                  V8             V9             V10               V11
#> 1              <NA>           <NA>            <NA>              <NA>
#> 2              <NA>           <NA>            <NA>              <NA>
#> 3              <NA>           <NA>            <NA>              <NA>
#> 4              <NA>           <NA>            <NA>              <NA>
#> 5              <NA>           <NA>            <NA>              <NA>
#> 6              <NA>           <NA>            <NA>              <NA>
#> 7              <NA>           <NA>            <NA>              <NA>
#> 8              <NA>           <NA>            <NA>              <NA>
#> 9              <NA>           <NA>            <NA>              <NA>
#> 10             <NA>           <NA>            <NA>              <NA>
#> 11             <NA>           <NA>            <NA>              <NA>
#> 12  social_facility     veterinary            <NA>              <NA>
#> 13             <NA>           <NA>            <NA>              <NA>
#> 14             <NA>           <NA>            <NA>              <NA>
#> 15             <NA>           <NA>            <NA>              <NA>
#> 16             <NA>           <NA>            <NA>              <NA>
#> 17           stable            sty            <NA>              <NA>
#> 18    plant_nursery           <NA>            <NA>              <NA>
#> 19             <NA>           <NA>            <NA>              <NA>
#> 20             <NA>           <NA>            <NA>              <NA>
#> 21             <NA>           <NA>            <NA>              <NA>
#> 22             <NA>           <NA>            <NA>              <NA>
#> 23             <NA>           <NA>            <NA>              <NA>
#> 24             <NA>           <NA>            <NA>              <NA>
#> 25             <NA>           <NA>            <NA>              <NA>
#> 26             <NA>           <NA>            <NA>              <NA>
#> 27             <NA>           <NA>            <NA>              <NA>
#> 28             <NA>           <NA>            <NA>              <NA>
#> 29             <NA>           <NA>            <NA>              <NA>
#> 30             <NA>           <NA>            <NA>              <NA>
#> 31             <NA>           <NA>            <NA>              <NA>
#> 32             <NA>           <NA>            <NA>              <NA>
#> 33             <NA>           <NA>            <NA>              <NA>
#> 34             <NA>           <NA>            <NA>              <NA>
#> 35             <NA>           <NA>            <NA>              <NA>
#> 36             <NA>           <NA>            <NA>              <NA>
#> 37       love_hotel      nightclub     planetarium         stripclub
#> 38             <NA>           <NA>            <NA>              <NA>
#> 39             <NA>           <NA>            <NA>              <NA>
#> 40             <NA>           <NA>            <NA>              <NA>
#> 41             <NA>           <NA>            <NA>              <NA>
#> 42   ranger_station       townhall public_building              <NA>
#> 43             <NA>           <NA>            <NA>              <NA>
#> 44             <NA>           <NA>            <NA>              <NA>
#> 45             <NA>           <NA>            <NA>              <NA>
#> 46             <NA>           <NA>            <NA>              <NA>
#> 47             <NA>           <NA>            <NA>              <NA>
#> 48             <NA>           <NA>            <NA>              <NA>
#> 49             <NA>           <NA>            <NA>              <NA>
#> 50             <NA>           <NA>            <NA>              <NA>
#> 51             <NA>           <NA>            <NA>              <NA>
#> 52             <NA>           <NA>            <NA>              <NA>
#> 53             <NA>           <NA>            <NA>              <NA>
#> 54       car_rental    car_sharing        car_wash    compressed_air
#> 55         services   speed_camera            <NA>              <NA>
#> 56             <NA>           <NA>            <NA>              <NA>
#> 57             <NA>           <NA>            <NA>              <NA>
#> 58             <NA>           <NA>            <NA>              <NA>
#> 59             <NA>           <NA>            <NA>              <NA>
#> 60             <NA>           <NA>            <NA>              <NA>
#> 61             <NA>           <NA>            <NA>              <NA>
#> 62             <NA>           <NA>            <NA>              <NA>
#> 63             <NA>           <NA>            <NA>              <NA>
#> 64             <NA>           <NA>            <NA>              <NA>
#> 65             <NA>           <NA>            <NA>              <NA>
#> 66             <NA>           <NA>            <NA>              <NA>
#> 67             <NA>           <NA>            <NA>              <NA>
#> 68             <NA>           <NA>            <NA>              <NA>
#> 69             <NA>           <NA>            <NA>              <NA>
#> 70             <NA>           <NA>            <NA>              <NA>
#> 71             <NA>           <NA>            <NA>              <NA>
#> 72             <NA>           <NA>            <NA>              <NA>
#> 73             <NA>           <NA>            <NA>              <NA>
#> 74           shrine      synagogue          temple         religious
#> 75             <NA>           <NA>            <NA>              <NA>
#> 76             <NA>           <NA>            <NA>              <NA>
#> 77             <NA>           <NA>            <NA>              <NA>
#> 78             <NA>           <NA>            <NA>              <NA>
#> 79             <NA>           <NA>            <NA>              <NA>
#> 80             <NA>           <NA>            <NA>              <NA>
#> 81      residential           <NA>            <NA>              <NA>
#> 82             <NA>           <NA>            <NA>              <NA>
#> 83             <NA>           <NA>            <NA>              <NA>
#> 84             <NA>           <NA>            <NA>              <NA>
#> 85             <NA>           <NA>            <NA>              <NA>
#> 86             <NA>           <NA>            <NA>              <NA>
#> 87             <NA>           <NA>            <NA>              <NA>
#> 88             <NA>           <NA>            <NA>              <NA>
#> 89             <NA>           <NA>            <NA>              <NA>
#> 90             <NA>           <NA>            <NA>              <NA>
#> 91             <NA>           <NA>            <NA>              <NA>
#> 92             <NA>           <NA>            <NA>              <NA>
#> 93             <NA>           <NA>            <NA>              <NA>
#> 94             <NA>           <NA>            <NA>              <NA>
#> 95             <NA>           <NA>            <NA>              <NA>
#> 96             <NA>           <NA>            <NA>              <NA>
#> 97             pump    pumping_rig pumping_station reservoir_covered
#> 98             <NA>           <NA>            <NA>              <NA>
#> 99             <NA>           <NA>            <NA>              <NA>
#> 100            <NA>           <NA>            <NA>              <NA>
#> 101            <NA>           <NA>            <NA>              <NA>
#> 102            <NA>           <NA>            <NA>              <NA>
#> 103  nature_reserve   picnic_table         fishing            garden
#> 104            <NA>           <NA>            <NA>              <NA>
#> 105            <NA>           <NA>            <NA>              <NA>
#> 106   dressing_room drinking_water        give_box     parcel_locker
#> 107            <NA>           <NA>            <NA>              <NA>
#> 108            <NA>           <NA>            <NA>              <NA>
#>                    V12              V13            V14         V15
#> 1                 <NA>             <NA>           <NA>        <NA>
#> 2                 <NA>             <NA>           <NA>        <NA>
#> 3                 <NA>             <NA>           <NA>        <NA>
#> 4                 <NA>             <NA>           <NA>        <NA>
#> 5                 <NA>             <NA>           <NA>        <NA>
#> 6                 <NA>             <NA>           <NA>        <NA>
#> 7                 <NA>             <NA>           <NA>        <NA>
#> 8                 <NA>             <NA>           <NA>        <NA>
#> 9                 <NA>             <NA>           <NA>        <NA>
#> 10                <NA>             <NA>           <NA>        <NA>
#> 11                <NA>             <NA>           <NA>        <NA>
#> 12                <NA>             <NA>           <NA>        <NA>
#> 13                <NA>             <NA>           <NA>        <NA>
#> 14                <NA>             <NA>           <NA>        <NA>
#> 15                <NA>             <NA>           <NA>        <NA>
#> 16                <NA>             <NA>           <NA>        <NA>
#> 17                <NA>             <NA>           <NA>        <NA>
#> 18                <NA>             <NA>           <NA>        <NA>
#> 19                <NA>             <NA>           <NA>        <NA>
#> 20                <NA>             <NA>           <NA>        <NA>
#> 21                <NA>             <NA>           <NA>        <NA>
#> 22                <NA>             <NA>           <NA>        <NA>
#> 23                <NA>             <NA>           <NA>        <NA>
#> 24                <NA>             <NA>           <NA>        <NA>
#> 25                <NA>             <NA>           <NA>        <NA>
#> 26                <NA>             <NA>           <NA>        <NA>
#> 27                <NA>             <NA>           <NA>        <NA>
#> 28                <NA>             <NA>           <NA>        <NA>
#> 29                <NA>             <NA>           <NA>        <NA>
#> 30                <NA>             <NA>           <NA>        <NA>
#> 31                <NA>             <NA>           <NA>        <NA>
#> 32                <NA>             <NA>           <NA>        <NA>
#> 33                <NA>             <NA>           <NA>        <NA>
#> 34                <NA>             <NA>           <NA>        <NA>
#> 35                <NA>             <NA>           <NA>        <NA>
#> 36                <NA>             <NA>           <NA>        <NA>
#> 37         swingerclub          theatre           <NA>        <NA>
#> 38                <NA>             <NA>           <NA>        <NA>
#> 39                <NA>             <NA>           <NA>        <NA>
#> 40                <NA>             <NA>           <NA>        <NA>
#> 41                <NA>             <NA>           <NA>        <NA>
#> 42                <NA>             <NA>           <NA>        <NA>
#> 43                <NA>             <NA>           <NA>        <NA>
#> 44                <NA>             <NA>           <NA>        <NA>
#> 45                <NA>             <NA>           <NA>        <NA>
#> 46                <NA>             <NA>           <NA>        <NA>
#> 47                <NA>             <NA>           <NA>        <NA>
#> 48                <NA>             <NA>           <NA>        <NA>
#> 49                <NA>             <NA>           <NA>        <NA>
#> 50                <NA>             <NA>           <NA>        <NA>
#> 51                <NA>             <NA>           <NA>        <NA>
#> 52                <NA>             <NA>           <NA>        <NA>
#> 53                <NA>             <NA>           <NA>        <NA>
#> 54  vehicle_inspection charging_station ferry_terminal        fuel
#> 55                <NA>             <NA>           <NA>        <NA>
#> 56                <NA>             <NA>           <NA>        <NA>
#> 57                <NA>             <NA>           <NA>        <NA>
#> 58                <NA>             <NA>           <NA>        <NA>
#> 59                <NA>             <NA>           <NA>        <NA>
#> 60                <NA>             <NA>           <NA>        <NA>
#> 61                <NA>             <NA>           <NA>        <NA>
#> 62                <NA>             <NA>           <NA>        <NA>
#> 63                <NA>             <NA>           <NA>        <NA>
#> 64                <NA>             <NA>           <NA>        <NA>
#> 65                <NA>             <NA>           <NA>        <NA>
#> 66                <NA>             <NA>           <NA>        <NA>
#> 67                <NA>             <NA>           <NA>        <NA>
#> 68                <NA>             <NA>           <NA>        <NA>
#> 69                <NA>             <NA>           <NA>        <NA>
#> 70                <NA>             <NA>           <NA>        <NA>
#> 71                <NA>             <NA>           <NA>        <NA>
#> 72                <NA>             <NA>           <NA>        <NA>
#> 73                <NA>             <NA>           <NA>        <NA>
#> 74                <NA>             <NA>           <NA>        <NA>
#> 75                <NA>             <NA>           <NA>        <NA>
#> 76                <NA>             <NA>           <NA>        <NA>
#> 77                <NA>             <NA>           <NA>        <NA>
#> 78                <NA>             <NA>           <NA>        <NA>
#> 79                <NA>             <NA>           <NA>        <NA>
#> 80                <NA>             <NA>           <NA>        <NA>
#> 81                <NA>             <NA>           <NA>        <NA>
#> 82                <NA>             <NA>           <NA>        <NA>
#> 83                <NA>             <NA>           <NA>        <NA>
#> 84                <NA>             <NA>           <NA>        <NA>
#> 85                <NA>             <NA>           <NA>        <NA>
#> 86                <NA>             <NA>           <NA>        <NA>
#> 87                <NA>             <NA>           <NA>        <NA>
#> 88                <NA>             <NA>           <NA>        <NA>
#> 89                <NA>             <NA>           <NA>        <NA>
#> 90                <NA>             <NA>           <NA>        <NA>
#> 91                <NA>             <NA>           <NA>        <NA>
#> 92                <NA>             <NA>           <NA>        <NA>
#> 93                <NA>             <NA>           <NA>        <NA>
#> 94                <NA>             <NA>           <NA>        <NA>
#> 95                <NA>             <NA>           <NA>        <NA>
#> 96                <NA>             <NA>           <NA>        <NA>
#> 97      street_cabinet             <NA>           <NA>        <NA>
#> 98                <NA>             <NA>           <NA>        <NA>
#> 99                <NA>             <NA>           <NA>        <NA>
#> 100               <NA>             <NA>           <NA>        <NA>
#> 101               <NA>             <NA>           <NA>        <NA>
#> 102               <NA>             <NA>           <NA>        <NA>
#> 103         playground    swimming_area           <NA>        <NA>
#> 104               <NA>             <NA>           <NA>        <NA>
#> 105               <NA>             <NA>           <NA>        <NA>
#> 106            shelter           shower        toilets water_point
#> 107               <NA>             <NA>           <NA>        <NA>
#> 108               <NA>             <NA>           <NA>        <NA>
#>                V16                V17             V18              V19
#> 1             <NA>               <NA>            <NA>             <NA>
#> 2             <NA>               <NA>            <NA>             <NA>
#> 3             <NA>               <NA>            <NA>             <NA>
#> 4             <NA>               <NA>            <NA>             <NA>
#> 5             <NA>               <NA>            <NA>             <NA>
#> 6             <NA>               <NA>            <NA>             <NA>
#> 7             <NA>               <NA>            <NA>             <NA>
#> 8             <NA>               <NA>            <NA>             <NA>
#> 9             <NA>               <NA>            <NA>             <NA>
#> 10            <NA>               <NA>            <NA>             <NA>
#> 11            <NA>               <NA>            <NA>             <NA>
#> 12            <NA>               <NA>            <NA>             <NA>
#> 13            <NA>               <NA>            <NA>             <NA>
#> 14            <NA>               <NA>            <NA>             <NA>
#> 15            <NA>               <NA>            <NA>             <NA>
#> 16            <NA>               <NA>            <NA>             <NA>
#> 17            <NA>               <NA>            <NA>             <NA>
#> 18            <NA>               <NA>            <NA>             <NA>
#> 19            <NA>               <NA>            <NA>             <NA>
#> 20            <NA>               <NA>            <NA>             <NA>
#> 21            <NA>               <NA>            <NA>             <NA>
#> 22            <NA>               <NA>            <NA>             <NA>
#> 23            <NA>               <NA>            <NA>             <NA>
#> 24            <NA>               <NA>            <NA>             <NA>
#> 25            <NA>               <NA>            <NA>             <NA>
#> 26            <NA>               <NA>            <NA>             <NA>
#> 27            <NA>               <NA>            <NA>             <NA>
#> 28            <NA>               <NA>            <NA>             <NA>
#> 29            <NA>               <NA>            <NA>             <NA>
#> 30            <NA>               <NA>            <NA>             <NA>
#> 31            <NA>               <NA>            <NA>             <NA>
#> 32            <NA>               <NA>            <NA>             <NA>
#> 33            <NA>               <NA>            <NA>             <NA>
#> 34            <NA>               <NA>            <NA>             <NA>
#> 35            <NA>               <NA>            <NA>             <NA>
#> 36            <NA>               <NA>            <NA>             <NA>
#> 37            <NA>               <NA>            <NA>             <NA>
#> 38            <NA>               <NA>            <NA>             <NA>
#> 39            <NA>               <NA>            <NA>             <NA>
#> 40            <NA>               <NA>            <NA>             <NA>
#> 41            <NA>               <NA>            <NA>             <NA>
#> 42            <NA>               <NA>            <NA>             <NA>
#> 43            <NA>               <NA>            <NA>             <NA>
#> 44            <NA>               <NA>            <NA>             <NA>
#> 45            <NA>               <NA>            <NA>             <NA>
#> 46            <NA>               <NA>            <NA>             <NA>
#> 47            <NA>               <NA>            <NA>             <NA>
#> 48            <NA>               <NA>            <NA>             <NA>
#> 49            <NA>               <NA>            <NA>             <NA>
#> 50            <NA>               <NA>            <NA>             <NA>
#> 51            <NA>               <NA>            <NA>             <NA>
#> 52            <NA>               <NA>            <NA>             <NA>
#> 53            <NA>               <NA>            <NA>             <NA>
#> 54        grit_bin motorcycle_parking         parking parking_entrance
#> 55            <NA>               <NA>            <NA>             <NA>
#> 56            <NA>               <NA>            <NA>             <NA>
#> 57            <NA>               <NA>            <NA>             <NA>
#> 58            <NA>               <NA>            <NA>             <NA>
#> 59            <NA>               <NA>            <NA>             <NA>
#> 60            <NA>               <NA>            <NA>             <NA>
#> 61            <NA>               <NA>            <NA>             <NA>
#> 62            <NA>               <NA>            <NA>             <NA>
#> 63            <NA>               <NA>            <NA>             <NA>
#> 64            <NA>               <NA>            <NA>             <NA>
#> 65            <NA>               <NA>            <NA>             <NA>
#> 66            <NA>               <NA>            <NA>             <NA>
#> 67            <NA>               <NA>            <NA>             <NA>
#> 68            <NA>               <NA>            <NA>             <NA>
#> 69            <NA>               <NA>            <NA>             <NA>
#> 70            <NA>               <NA>            <NA>             <NA>
#> 71            <NA>               <NA>            <NA>             <NA>
#> 72            <NA>               <NA>            <NA>             <NA>
#> 73            <NA>               <NA>            <NA>             <NA>
#> 74            <NA>               <NA>            <NA>             <NA>
#> 75            <NA>               <NA>            <NA>             <NA>
#> 76            <NA>               <NA>            <NA>             <NA>
#> 77            <NA>               <NA>            <NA>             <NA>
#> 78            <NA>               <NA>            <NA>             <NA>
#> 79            <NA>               <NA>            <NA>             <NA>
#> 80            <NA>               <NA>            <NA>             <NA>
#> 81            <NA>               <NA>            <NA>             <NA>
#> 82            <NA>               <NA>            <NA>             <NA>
#> 83            <NA>               <NA>            <NA>             <NA>
#> 84            <NA>               <NA>            <NA>             <NA>
#> 85            <NA>               <NA>            <NA>             <NA>
#> 86            <NA>               <NA>            <NA>             <NA>
#> 87            <NA>               <NA>            <NA>             <NA>
#> 88            <NA>               <NA>            <NA>             <NA>
#> 89            <NA>               <NA>            <NA>             <NA>
#> 90            <NA>               <NA>            <NA>             <NA>
#> 91            <NA>               <NA>            <NA>             <NA>
#> 92            <NA>               <NA>            <NA>             <NA>
#> 93            <NA>               <NA>            <NA>             <NA>
#> 94            <NA>               <NA>            <NA>             <NA>
#> 95            <NA>               <NA>            <NA>             <NA>
#> 96            <NA>               <NA>            <NA>             <NA>
#> 97            <NA>               <NA>            <NA>             <NA>
#> 98            <NA>               <NA>            <NA>             <NA>
#> 99            <NA>               <NA>            <NA>             <NA>
#> 100           <NA>               <NA>            <NA>             <NA>
#> 101           <NA>               <NA>            <NA>             <NA>
#> 102           <NA>               <NA>            <NA>             <NA>
#> 103           <NA>               <NA>            <NA>             <NA>
#> 104           <NA>               <NA>            <NA>             <NA>
#> 105           <NA>               <NA>            <NA>             <NA>
#> 106 watering_place       refugee_site vending_machine             <NA>
#> 107           <NA>               <NA>            <NA>             <NA>
#> 108           <NA>               <NA>            <NA>             <NA>
#>               V20  V21
#> 1            <NA> <NA>
#> 2            <NA> <NA>
#> 3            <NA> <NA>
#> 4            <NA> <NA>
#> 5            <NA> <NA>
#> 6            <NA> <NA>
#> 7            <NA> <NA>
#> 8            <NA> <NA>
#> 9            <NA> <NA>
#> 10           <NA> <NA>
#> 11           <NA> <NA>
#> 12           <NA> <NA>
#> 13           <NA> <NA>
#> 14           <NA> <NA>
#> 15           <NA> <NA>
#> 16           <NA> <NA>
#> 17           <NA> <NA>
#> 18           <NA> <NA>
#> 19           <NA> <NA>
#> 20           <NA> <NA>
#> 21           <NA> <NA>
#> 22           <NA> <NA>
#> 23           <NA> <NA>
#> 24           <NA> <NA>
#> 25           <NA> <NA>
#> 26           <NA> <NA>
#> 27           <NA> <NA>
#> 28           <NA> <NA>
#> 29           <NA> <NA>
#> 30           <NA> <NA>
#> 31           <NA> <NA>
#> 32           <NA> <NA>
#> 33           <NA> <NA>
#> 34           <NA> <NA>
#> 35           <NA> <NA>
#> 36           <NA> <NA>
#> 37           <NA> <NA>
#> 38           <NA> <NA>
#> 39           <NA> <NA>
#> 40           <NA> <NA>
#> 41           <NA> <NA>
#> 42           <NA> <NA>
#> 43           <NA> <NA>
#> 44           <NA> <NA>
#> 45           <NA> <NA>
#> 46           <NA> <NA>
#> 47           <NA> <NA>
#> 48           <NA> <NA>
#> 49           <NA> <NA>
#> 50           <NA> <NA>
#> 51           <NA> <NA>
#> 52           <NA> <NA>
#> 53           <NA> <NA>
#> 54  parking_space taxi
#> 55           <NA> <NA>
#> 56           <NA> <NA>
#> 57           <NA> <NA>
#> 58           <NA> <NA>
#> 59           <NA> <NA>
#> 60           <NA> <NA>
#> 61           <NA> <NA>
#> 62           <NA> <NA>
#> 63           <NA> <NA>
#> 64           <NA> <NA>
#> 65           <NA> <NA>
#> 66           <NA> <NA>
#> 67           <NA> <NA>
#> 68           <NA> <NA>
#> 69           <NA> <NA>
#> 70           <NA> <NA>
#> 71           <NA> <NA>
#> 72           <NA> <NA>
#> 73           <NA> <NA>
#> 74           <NA> <NA>
#> 75           <NA> <NA>
#> 76           <NA> <NA>
#> 77           <NA> <NA>
#> 78           <NA> <NA>
#> 79           <NA> <NA>
#> 80           <NA> <NA>
#> 81           <NA> <NA>
#> 82           <NA> <NA>
#> 83           <NA> <NA>
#> 84           <NA> <NA>
#> 85           <NA> <NA>
#> 86           <NA> <NA>
#> 87           <NA> <NA>
#> 88           <NA> <NA>
#> 89           <NA> <NA>
#> 90           <NA> <NA>
#> 91           <NA> <NA>
#> 92           <NA> <NA>
#> 93           <NA> <NA>
#> 94           <NA> <NA>
#> 95           <NA> <NA>
#> 96           <NA> <NA>
#> 97           <NA> <NA>
#> 98           <NA> <NA>
#> 99           <NA> <NA>
#> 100          <NA> <NA>
#> 101          <NA> <NA>
#> 102          <NA> <NA>
#> 103          <NA> <NA>
#> 104          <NA> <NA>
#> 105          <NA> <NA>
#> 106          <NA> <NA>
#> 107          <NA> <NA>
#> 108          <NA> <NA>
collapse::unlist2d(osm_line_class, idcols = c("category", "tag"))
#>     category           tag            V1                V2
#> 1  aerialway     aerialway                            <NA>
#> 2    aeroway       aeroway                            <NA>
#> 3   pipeline      man_made      pipeline              <NA>
#> 4       road       highway      motorway     motorway_link
#> 5    railway       railway                            <NA>
#> 6   waterway      waterway    !waterfall           !stream
#> 7   waterway         water          !yes              <NA>
#> 8   waterway      man_made   water_works reservoir_covered
#> 9      power         power                            <NA>
#> 10   telecom       telecom                            <NA>
#> 11   telecom communication                            <NA>
#> 12   storage      man_made          silo      storage_tank
#> 13  boundary      boundary national_park    protected_area
#> 14     ferry         route         ferry              <NA>
#>                       V3         V4      V5             V6         V7
#> 1                   <NA>       <NA>    <NA>           <NA>       <NA>
#> 2                   <NA>       <NA>    <NA>           <NA>       <NA>
#> 3                   <NA>       <NA>    <NA>           <NA>       <NA>
#> 4                  trunk trunk_link primary   primary_link  secondary
#> 5                   <NA>       <NA>    <NA>           <NA>       <NA>
#> 6            !stream_end      !wadi !rapids !tidal_channel !riverbank
#> 7                   <NA>       <NA>    <NA>           <NA>       <NA>
#> 8            water_tower       <NA>    <NA>           <NA>       <NA>
#> 9                   <NA>       <NA>    <NA>           <NA>       <NA>
#> 10                  <NA>       <NA>    <NA>           <NA>       <NA>
#> 11                  <NA>       <NA>    <NA>           <NA>       <NA>
#> 12                  <NA>       <NA>    <NA>           <NA>       <NA>
#> 13 special_economic_zone       <NA>    <NA>           <NA>       <NA>
#> 14                  <NA>       <NA>    <NA>           <NA>       <NA>
#>                V8       V9           V10
#> 1            <NA>     <NA>          <NA>
#> 2            <NA>     <NA>          <NA>
#> 3            <NA>     <NA>          <NA>
#> 4  secondary_link tertiary tertiary_link
#> 5            <NA>     <NA>          <NA>
#> 6            !yes     <NA>          <NA>
#> 7            <NA>     <NA>          <NA>
#> 8            <NA>     <NA>          <NA>
#> 9            <NA>     <NA>          <NA>
#> 10           <NA>     <NA>          <NA>
#> 11           <NA>     <NA>          <NA>
#> 12           <NA>     <NA>          <NA>
#> 13           <NA>     <NA>          <NA>
#> 14           <NA>     <NA>          <NA>
# This list contains additional tags with information about lines (e.g. roads and railways)
collapse::unlist2d(osm_line_info_tags, idcols = c("category", "tag"))
#>     category   V1  V2   V3          V4       V5    V6      V7       V8       V9
#> 1  aerialway type ref name description operator usage service capacity man_made
#> 2    aeroway type ref name description operator usage service capacity man_made
#> 3   pipeline type ref name description operator usage service capacity man_made
#> 4       road type ref name description operator usage service capacity man_made
#> 5    railway type ref name description operator usage service capacity man_made
#> 6   waterway type ref name description operator usage service capacity man_made
#> 7      power type ref name description operator usage service capacity man_made
#> 8    telecom type ref name description operator usage service capacity man_made
#> 9    storage type ref name description operator usage service capacity man_made
#> 10  boundary type ref name description operator usage service capacity man_made
#> 11     route type ref name description operator usage service capacity man_made
#>             V10            V11              V12              V13            V14
#> 1     aerialway      occupancy         duration       detachable         bubble
#> 2     aerialway        surface              lit    navigationaid           iata
#> 3      pipeline          depth            inlet           outlet   construction
#> 4       highway       abutters            lanes              lit       maxspeed
#> 5       railway         tracks      electrified       embankment embedded_rails
#> 6  intermittent          inlet           outlet            width          draft
#> 7         power    electricity     plant:output generator:output         length
#> 8       telecom telecom:medium connection_point          utility         length
#> 9       content           crop           height         material         length
#> 10  related_law  protect_class protection_title    official_name         length
#> 11       access       duration    opening_hours      reservation  motor_vehicle
#>                V15           V16        V17          V18         V19
#> 1          heating       bicycle     access       length       width
#> 2             icao       amenity    highway       length       width
#> 3        substance       content   resource     diameter    pressure
#> 4  traffic_calming        oneway    surface   smoothness   tracktype
#> 5            gauge loading_gauge metre_load    axle_load     voltage
#> 6            tidal      maxwidth  maxheight    maxlength    maxspeed
#> 7            width      location     origin place:origin destination
#> 8            width      location     origin place:origin destination
#> 9            width      location     origin place:origin destination
#> 10           width      location     origin place:origin destination
#> 11            foot     maxweight  maxlength     maxwidth   maxheight
#>                  V20        V21          V22          V23               V24
#> 1           location     origin place:origin  destination place:destination
#> 2           location     origin place:origin  destination place:destination
#> 3     flow_direction      count       length        width          location
#> 4             length      width     location       origin      place:origin
#> 5          frequency  highspeed     maxspeed     historic            length
#> 6             length   location       origin place:origin       destination
#> 7  place:destination start_date     end_date         <NA>              <NA>
#> 8  place:destination start_date     end_date         <NA>              <NA>
#> 9  place:destination start_date     end_date         <NA>              <NA>
#> 10 place:destination start_date     end_date         <NA>              <NA>
#> 11            length      width     location       origin      place:origin
#>                  V25               V26         V27               V28
#> 1         start_date          end_date        <NA>              <NA>
#> 2         start_date          end_date        <NA>              <NA>
#> 3             origin      place:origin destination place:destination
#> 4        destination place:destination  start_date          end_date
#> 5              width          location      origin      place:origin
#> 6  place:destination        start_date    end_date              <NA>
#> 7               <NA>              <NA>        <NA>              <NA>
#> 8               <NA>              <NA>        <NA>              <NA>
#> 9               <NA>              <NA>        <NA>              <NA>
#> 10              <NA>              <NA>        <NA>              <NA>
#> 11       destination place:destination  start_date          end_date
#>            V29               V30        V31      V32
#> 1         <NA>              <NA>       <NA>     <NA>
#> 2         <NA>              <NA>       <NA>     <NA>
#> 3   start_date          end_date       <NA>     <NA>
#> 4         <NA>              <NA>       <NA>     <NA>
#> 5  destination place:destination start_date end_date
#> 6         <NA>              <NA>       <NA>     <NA>
#> 7         <NA>              <NA>       <NA>     <NA>
#> 8         <NA>              <NA>       <NA>     <NA>
#> 9         <NA>              <NA>       <NA>     <NA>
#> 10        <NA>              <NA>       <NA>     <NA>
#> 11        <NA>              <NA>       <NA>     <NA>